requirements.txt
aiosqlite==0.22.1
anthropic==0.83.0
httpx==0.28.1
py-clob-client==0.34.6
python-dotenv==1.2.1
python-telegram-bot==22.6
requests==2.32.5
scipy==1.17.0
